# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Vehicle {
  id: number;
  slug: string;
  year: number;
  make: string;
  model: string;
  price: number;
  mileage: number;
  location: string;
  bodyStyle: string;
  transmission: string;
  engine: string;
  imageUrl: string;
  description: string;
  featured: boolean;
}

export interface SoldVehicle {
  id: number;
  year: number;
  make: string;
  model: string;
  soldDate: string;
  imageUrl: string;
  location: string;
  /** @nullable */
  mileageKm?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  galleryUrls?: string[] | null;
}

export interface ImportService {
  id: number;
  name: string;
  description: string;
  details: string[];
}

export interface SiteSummary {
  availableCount: number;
  soldCount: number;
  yearsExperience: number;
  statesCovered: number;
}

export type InquiryInputInquiryType = typeof InquiryInputInquiryType[keyof typeof InquiryInputInquiryType];


export const InquiryInputInquiryType = {
  vehicle: 'vehicle',
  transport: 'transport',
  general: 'general',
} as const;

export interface InquiryInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  vehicleSlug?: string | null;
  inquiryType: InquiryInputInquiryType;
  /** @minLength 10 */
  message: string;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  vehicleSlug?: string | null;
  inquiryType: string;
  message: string;
  createdAt: string;
}

export type ListInventoryParams = {
search?: string;
bodyStyle?: string;
sort?: ListInventorySort;
};

export type ListInventorySort = typeof ListInventorySort[keyof typeof ListInventorySort];


export const ListInventorySort = {
  featured: 'featured',
  newest: 'newest',
  'price-low': 'price-low',
  'price-high': 'price-high',
} as const;

export type ListSoldVehiclesParams = {
search?: string;
};