# lib/api-zod/src/generated/types/soldVehicle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SoldVehicle {
  id: number;
  year: number;
  make: string;
  model: string;
  soldDate: string;
  imageUrl: string;
  location: string;
  /** @nullable */
  mileageKm?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  galleryUrls?: string[] | null;
}